Removed long longs

+++ src/editor/objects/navmesh.cpp
@@ -6,9 +6,9 @@ namespace Editor {
 
 using namespace tram;
 
-static uint64_t navmesh_id_counter = 0;
+static uint32_t navmesh_id_counter = 0;
 
-uint64_t Navmesh::GetNewID() {
+uint32_t Navmesh::GetNewID() {
     return ++navmesh_id_counter;
 }
 
@@ -30,7 +30,7 @@ void Navmesh::LoadFromDisk() {
     file.read_name(); // skip navmesh name
     
     while (file.is_continue()) {
-        uint64_t node_id = file.read_uint32();
+        uint32_t node_id = file.read_uint32();
         
         if (node_id > navmesh_id_counter) {
             navmesh_id_counter = node_id;
@@ -40,10 +40,10 @@ void Navmesh::LoadFromDisk() {
         
         new_node->SetProperty("id", node_id);
         
-        new_node->SetProperty("next-id", (uint64_t)file.read_uint32());
-        new_node->SetProperty("prev-id", (uint64_t)file.read_uint32());
-        new_node->SetProperty("left-id", (uint64_t)file.read_uint32());
-        new_node->SetProperty("right-id", (uint64_t)file.read_uint32());
+        new_node->SetProperty("next-id", file.read_uint32());
+        new_node->SetProperty("prev-id", file.read_uint32());
+        new_node->SetProperty("left-id", file.read_uint32());
+        new_node->SetProperty("right-id", file.read_uint32());
         
         new_node->SetProperty("position-x", file.read_float32());
         new_node->SetProperty("position-y", file.read_float32());
